refactor : adapter to temporal
